change authKeyId from BigInteger to long

+++ src/main/java/com/im/lib/net/DispatcherWebsocket.java
@@ -19,7 +19,6 @@ import org.springframework.stereotype.Component;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -97,7 +96,7 @@ public class DispatcherWebsocket implements ApplicationContextAware {
                     + "方法未实现！");
         }
 
-        BigInteger authKeyId = requestData.getAuthKeyId();
+        long authKeyId = requestData.getAuthKeyId();
         return RpcResult.ok(authKeyId, response, requestData.getSessionId());
     }
